convert(String, ::Vector{Char}): don't normalize surrogate pairs
we're in a new era where Julia tries as hard as it can to pass text
data through intact rather than trying to guess what you meant. to
that end, when converting a vector of characters to a `String` we
no longer detect and combine surrogate pairs. there can be a utility
function to do that kind of normalization on strings, but it's not
the job of the basic string code.